Fix entity indexing
Components without explicit comparison functions
were being considered equal by default, which was
causing some components to be ignored when adding
to the indices. Entity IDs were also not being
generated correctly.

It may be necessary in the future to allow for
single components to be associated with multiple
entities.